add exoplayer player layout, added onClick listeners to app navigation buttons that change fragments
