+++ app/src/main/java/com/example/tacomamusicplayer/activity/ChooseMusicActivity.kt
@@ -3,12 +3,14 @@ package com.example.tacomamusicplayer.activity
 import android.os.Build
 import android.os.Bundle
 import android.view.View
+import android.view.Window
 import android.view.WindowInsets
 import androidx.annotation.RequiresApi
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.widget.ViewPager2
 import com.example.tacomamusicplayer.adapter.ScreenSlidePagerAdapter
 import com.example.tacomamusicplayer.databinding.ActivityChooseMusicBinding
+import com.example.tacomamusicplayer.util.UtilImpl
 
 
 class ChooseMusicActivity: FragmentActivity() {
@@ -30,12 +32,25 @@ class ChooseMusicActivity: FragmentActivity() {
         binding = ActivityChooseMusicBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-
         //Instantiate a ViewPager2 and a PagerAdapter
         viewPager = binding.pager
         val pagerAdapter = ScreenSlidePagerAdapter(this)
         viewPager.adapter = pagerAdapter
 
+        binding.navigationControl.setPlaylistButtonOnClick {
+            viewPager.currentItem = 0
+        }
+        binding.navigationControl.setBrowseAlbumButtonOnClick {
+            viewPager.currentItem = 1
+        }
+        binding.navigationControl.setAlbumButtonOnClick {
+            viewPager.currentItem = 2
+        }
+    }
+
+    override fun onResume() {
+        super.onResume()
+        UtilImpl.hideNavigationUI(window)
     }
 
     override fun onBackPressed() {
@@ -46,23 +61,4 @@ class ChooseMusicActivity: FragmentActivity() {
             viewPager.currentItem = viewPager.currentItem - 1
         }
     }
-
-    /**
-     * Call this function onResume(). Removes the navigation bar from the bottom of the screen.
-     */
-    fun hideNavigationUI() {
-        if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            window.insetsController?.hide(WindowInsets.Type.navigationBars())
-        } else {
-            val flags = (View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-                    or View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                    or View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                    or View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
-                    or View.SYSTEM_UI_FLAG_FULLSCREEN
-                    or View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY)
-
-            window.decorView.systemUiVisibility = flags
-        }
-    }
-
 }
