+++ app/src/main/java/com/example/tacomamusicplayer/util/UtilImpl.kt
@@ -1,5 +1,9 @@
 package com.example.tacomamusicplayer.util
 
+import android.os.Build
+import android.view.View
+import android.view.Window
+import android.view.WindowInsets
 import kotlin.time.DurationUnit
 import kotlin.time.toDuration
 
@@ -15,6 +19,24 @@ class UtilImpl {
             val seconds = duration.minus(minutes.toDuration(DurationUnit.MINUTES)).inWholeSeconds
             //Return formatted string
             return "$minutes:$seconds"
+        }
+
+        /**
+         * Call this function onResume(). Removes the navigation bar from the bottom of the screen.
+         */
+        fun hideNavigationUI(window: Window) {
+            if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+                window.insetsController?.hide(WindowInsets.Type.navigationBars())
+            } else {
+                val flags = (View.SYSTEM_UI_FLAG_LAYOUT_STABLE
+                        or View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
+                        or View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
+                        or View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
+                        or View.SYSTEM_UI_FLAG_FULLSCREEN
+                        or View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY)
+
+                window.decorView.systemUiVisibility = flags
+            }
         }
     }
 }
