+++ app/src/main/res/layout/activity_main.xml
@@ -12,14 +12,15 @@
     <androidx.media3.ui.PlayerView
         android:id="@+id/player_view"
         android:layout_width="300dp"
-        android:layout_height="300dp"
+        android:layout_height="500dp"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:show_timeout="0"
         app:controller_layout_id="@layout/custom_exo_controller_update"
-        android:layout_marginTop="25dp"
-        android:background="#BA8484"
+        app:player_layout_id="@layout/custom_exo_player_update"
+        android:layout_marginTop="60dp"
+        android:background="@color/primary"
         app:hide_on_touch="false"
         app:use_controller="true"
         app:auto_show="false"
