+++ app/src/main/res/layout/custom_exo_controller_update.xml
@@ -5,7 +5,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_gravity="bottom"
+    android:layout_gravity="top"
     android:background="#00FFFFFF">
 
     <View android:id="@id/exo_progress_placeholder"
@@ -45,12 +45,6 @@
         android:layout_marginTop="35dp"
         android:layout_marginEnd="25dp"/>
 
-
-    <!--        <View-->
-    <!--            android:layout_height="1dp"-->
-    <!--            android:layout_weight="1"-->
-    <!--             android:layout_width="1dp"/>-->
-
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -61,14 +55,14 @@
 
         <ImageButton android:id="@id/exo_prev"
             style="@style/ExoMediaButton.Previous"
-            android:layout_marginTop="35dp"
+            android:layout_marginTop="25dp"
             android:layout_marginEnd="45dp"
             android:layout_weight="1"
             android:layout_marginLeft="100dp"/>
 
         <ImageButton android:id="@id/exo_play_pause"
             style="@style/ExoMediaButton.Play"
-            android:layout_marginTop="35dp"
+            android:layout_marginTop="25dp"
             android:layout_marginStart="20dp"
             android:layout_weight="1"
             android:layout_marginBottom="100dp"
@@ -76,7 +70,7 @@
 
         <ImageButton android:id="@id/exo_next"
             style="@style/ExoMediaButton.Next"
-            android:layout_marginTop="35dp"
+            android:layout_marginTop="25dp"
             android:layout_marginStart="45dp"
             android:layout_weight="1"
             android:layout_marginRight="100dp"/>
