+++ app/src/main/res/layout/custom_exo_player_update.xml
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="top"
+    android:orientation="vertical"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+<!--    <View android:id="@id/exo_shutter"-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="match_parent"-->
+<!--        android:background="@android:color/black"/>-->
+
+    <ImageView android:id="@id/exo_artwork"
+        android:layout_width="300dp"
+        android:layout_height="300dp"
+        android:scaleType="fitXY"/>
+
+    <View android:id="@id/exo_controller_placeholder"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_marginTop="30dp"/>
+
+<!--    <androidx.media3.ui.SubtitleView android:id="@id/exo_subtitles"-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="match_parent"/>-->
+
+<!--    <ProgressBar android:id="@id/exo_buffering"-->
+<!--        android:layout_width="wrap_content"-->
+<!--        android:layout_height="wrap_content"-->
+<!--        android:indeterminate="true"-->
+<!--        android:layout_gravity="center"/>-->
+
+<!--    <TextView android:id="@id/exo_error_message"-->
+<!--        android:layout_width="wrap_content"-->
+<!--        android:layout_height="@dimen/exo_error_message_height"-->
+<!--        android:layout_gravity="center"-->
+<!--        android:layout_marginBottom="@dimen/exo_error_message_margin_bottom"-->
+<!--        android:gravity="center"-->
+<!--        android:textColor="@color/exo_white"-->
+<!--        android:textSize="@dimen/exo_error_message_text_size"-->
+<!--        android:background="@drawable/exo_rounded_rectangle"-->
+<!--        android:paddingLeft="@dimen/exo_error_message_text_padding_horizontal"-->
+<!--        android:paddingRight="@dimen/exo_error_message_text_padding_horizontal"-->
+<!--        android:paddingTop="@dimen/exo_error_message_text_padding_vertical"-->
+<!--        android:paddingBottom="@dimen/exo_error_message_text_padding_vertical"/>-->
+<!--    -->
+
+<!--    <FrameLayout android:id="@id/exo_ad_overlay"-->
+<!--    android:layout_width="match_parent"-->
+<!--    android:layout_height="match_parent"/>-->
+<!--    -->
+<!--    <FrameLayout android:id="@id/exo_overlay"-->
+<!--    android:layout_width="match_parent"-->
+<!--    android:layout_height="match_parent"/>-->
+<!--    -->
+    </LinearLayout>
